More flexible header parsing.
In order to support more sophisticated header parsing, I am adding the ability
to customize the response code associated with the Error parsing response. In
particular, if the header is malformed for whatever reason, it makes more sense
to return a BAD_REQUEST than a FORBIDDEN response code.

Additionally, I am removing the unnecessary NoRestrictions parser.