Add combinators to StructuredAccessControl
To provide the maximal flexibility for end users, StructuredAccessControl
objects should be combine-able in a variety of ways in order to reasonably
define new policies.

While it is recommended for common constraints to use an AuthenticationCombiner
to smoothly define a new access control policy, there are a number of cases were
a more ad-hoc approach is justified.

In order to maintain composibility, we ensure to always return another
StructuredAccessControl. Note: in order to get appropriate semantics, in some
combinators, we need to run the Authorizer within the authenticator itself.

Because this is very important code, I've included a comprehensive test suite to
ensure the correct operation of the combinators.