Add testing capabilities to HeaderAccessControl (#33)
When testing Naptime endpoints, it's often a good idea to test the access
control mechanisms as well. Because Naptime tests operate at a higher level,
it doesn't make sense to include the parsing & decorating. It does, however,
make sense to run the Authorizors.

In order to do this while simultaneously supporting the combinators, we add a
new method to `HeaderAccessControl` called `check` that will run the Authorizer.
The type signature of `check` is a little funny in order to support the behavior
of some of the more sophisticated combinators (i.e. `anyOf`) which must support
filtering of the parsed values.