Incorporate related resource information in schema
Summary:
In order to build the automatic resource inclusion engine, to support
GraphQL, and add links to the documentation between related resources,
we need to incorporate the related fields information into the "merged"
type schema.

Because the Engine, GraphQL and others are centered around Courier and
Pegasus schemas, I've begun constructing a set of classes that can form
the basis for a more integration-style test suite in the
`NestedMacroCourierTests` class. The first test in this class is a test
of the related fields information functionality added in this commit.

Because we are centering upon courier models, I've added a couple extra
abstract classes that reduce the amount of boilerplate required for
resources using courier types.