Design response model + EngineImpl improvements
This commit implements the long term response model, as per the design
discussions. It converts all ARI-related code to use it, and adds a
number of tests to ensure correctness.

Additionally, thanks to the new response model, we are able to implement
a more sophisticated handling of multiple top-level requests in the
engine.