Infer additional key type schemas (#98)
It is quite common to use a more sophisticated key type. Often, these key types
have a single apply method in the companion object that corresopnds to the
fields of the serialized form of the type. This commit extends the macro to
automatically infer the schema for these more sophisticated key types.

Additionally, we capture the changes in some unit tests to verify the correct
operation of the macro.