Handle Key types that define their own id field (#102)
In some cases, a key courier type might define its own id field (e.g.
only a single part of the key is required to uniquely identify the
record). This commit makes the Types computation a little bit smarter.

Note that relying on this behavior is a potential code smell.
Conflicts:
	version.sbt