Release version 0.3.0
Naptime 0.3.0 is the first ARI+GraphQL release. It includes an initial
GraphQL implementation including the appropriate engine abstractions. As
part of the release, we include the automatic resource includes engine
implementation. Although it is a relatively naive implementation
currently, it should form the basis of future work.

Note: The 0.3.0 abstractions were developed on the `graphql` branch, and
have been rebased onto the current master at the time of this commit.