Rebase + release ARI+GraphQL (0.3.0) (#105)
* Proposal automatic resource inclusion abstractions (#64)

Define key abstractions that will form the basis for the automatic
resource inclusion and GraphQL work.

* Add initial GraphQL parsing library (#69)

- Adds sangria dependency to Naptime
- Adds a GraphQL Parser object that can parse a GraphQL String into a Naptime Request object
- Recursively parses fields into Naptime RequestFields
- Adds unit tests for parsing
- Add license, parse more types, and minor cleanups

Still a few TODOs left around error handling and argument parsing (GraphQL AST types into JSON)

* Add schema generation for GraphQL (#75)

Creates a helper class to convert a set of Naptime resources and Pegasus schemas to a [Sangria] GraphQL Schema

* Add data schemas to ari API. (#74)

The engine must also expose the set of types that make up the
collection of resources to the presentation layer.

* ARI: Initial Engine Impl Tests (#76)

Set up the basic infrastructure for testing the engine implementation.
Currently, we only have a trivial engine implementation that cannot
handle any automatic joining.

* Initial local fetcher implementation. (#77)

Implement the initial version of a local fetcher which an engine can
use to execute requests against local naptime resources.

* Add data-fetching to Sangria schemas and hook up to EngineApi (#78)

Updates the sangria schemas to pull fields and related resources from the Naptime engine response. Also adds a GraphQL controller and sample resources to the example application for testing.

* Design response model + EngineImpl improvements (#79)

This commit implements the long term response model, as per the design
discussions. It converts all ARI-related code to use it, and adds a
number of tests to ensure correctness.

Additionally, thanks to the new response model, we are able to implement
a more sophisticated handling of multiple top-level requests in the
engine.

* Support basic joins. (#80)

This commit adds support for a single level of automated nested/joining
of related resources. Subsequent commits will add support for more
foreign-key types of joins, as well as support alternate joins such
as reverse includes / etc.

Additionally, this commit cleans up the engine API. There were unused
components of the engine API that won't be used. By removing the
unnecessary additional fields in the engine API, we avoid tangling the
code and subsequent developer confusion.

* Support single element joins. (#81)

In this commit, we add support for joining against single foreign key
fields.

Additionally, we add a test to ensure that complex fields do not cause
the engine trouble.

* Add multi-join test. (#82)

Add a test to verify the engine correctly handles joining multiple
resources on to a top level resource.

Additionally, this commit reorganizes all the TODOs into a single list
in the EngineImplTest class.

* Fix computing full path. (#84)

Computing the full path is required to display the correct documentation.
In order to handle nested resources correctly, we need to stop recursing
when we hit the root resource. This commit fixes that.

* Support arbitrary recursive nesting of resources. (#85)

If Resource A points to Resource B, and Resource B points to Resource C
a single query should be able to link everything up. This commit extends
our engine from supporting only a single hop, to supporting multiple
hops.

* Update macro to support primitive defaults and misc. fixes (#87)

Updates the naptime resource macro to support defaults of a primitive type (instead of just record types), and wraps some potentially dangerous operations in schema building in a try / catch until @saeta gets schema models in a better state

* Clean up some code (#88)

* Separate graphql/sangria dependencies to own lib (#91)

In order to avoid polluting the classpath for naptime users that are not
interested in sangria and/or graphql, we have split up those
dependencies into a separate library called `naptime-graphql`.

* Release RC1 of Naptime 0.3.0 (#92)

* Handle more edge cases in schema generation (#89)

* Prepare to release version 0.3.0-RC2 (#93)

* Release RC3 for graphql. (#94)

* Handle recursive model definitions and null field definitions (#97)

* Abstract NaptimeRoutes behind SchemaProvider trait. (#104)

Previously, the EngineImpl directly pulled from the local naptime resources.
However, this does not work in distributed mode. Instead, we must add a layer
of indirection to allow for custom distributed / remote SchemaProviders.

* Set default fields for empty object types, and handle non-related record types (#103)

Additionally, we made a change to correctly parse GraphQL queries into the appropriate intermediate representation expected by the Engine implementation.

* Release version 0.3.0

Naptime 0.3.0 is the first ARI+GraphQL release. It includes an initial
GraphQL implementation including the appropriate engine abstractions. As
part of the release, we include the automatic resource includes engine
implementation. Although it is a relatively naive implementation
currently, it should form the basis of future work.

Note: The 0.3.0 abstractions were developed on the `graphql` branch, and
have been rebased onto the current master at the time of this commit.