Add DataMap support to JSON serializer and schemas
Some of our Courier models support a `passthroughExempt` flag, which lets the model support an arbitrary, untyped DataMap as a field. GraphQL/Sangria was freaking out when seeing this though, as there were no fields defined (which is not allowed according to GraphQL’s spec). This takes a lot of inspiration from both https://github.com/taion/graphql-type-json and https://gist.github.com/OlegIlyenko/5b96f4b54f656aac226d3c4bc33fd2a6 to add a `DataMapType` to the GraphQL schema we support.

TODO: add tests for this :). But initial thoughts would be nice, in case I have to end up refactoring any of this.