Properly “fix-up” data being parsed into DataMaps
Courier serializes Ints and Longs in the same format over the wire. When we parse the results into DataMaps, some Longs can be parsed and stored as Ints, which then leads to CastClassExceptions when trying to read the value as a Long. This change uses the expected DataSchema type and the “validate” method on DataMaps to correctly parse the value into the proper format.