Support query batching in the naptime controller
Note: this changes the way that the filters work to instead pass around JsObjects instead of a response. This means that filters / the GraphQL engine cannot alter the response at all (including response codes + headers), but instead must put all data on the response body and _always_ return a 200 access code. This is acceptable with the way that GraphQL works, but is a change to keep in mind