Fix bugs with generating and resolving union types (#143)
This fixes two bugs with union types, both around the way that names are converted from a courier name (i.e. org.coursera.UnionMember) to a GraphQL compatible name (org_coursera_UnionMember).

The first bug was that we could not correctly resolve the ObjectType to parse into based on the field names, because we were comparing strings of the two different formats, which were never matching. This wasn't caught when we were working with primitive types, because "int" didn't have any field name conversion issues.

The second bug was around how we were parsing the data out of the record field, which again was due to a mismatch in the two names.

These issues should be fixed now, and I added a unit test to check against more complex union types.