Add a metrics collector to pull execution time + parallelize batching (#157)
This adds a metrics collector that will collect timing data from the GraphQL controller. In Naptime, a logging collector is used to send the metrics as log statements. However, in other codebases, the data can be sent to a third party collector (i.e. Datadog).

This also makes a small change to better parallelize the GraphQL Batch endpoint