Support defining relations on nested models / unions
This change allows a relation (either forward or reverse), defined on the `Fields` object in a resource, to reference deeply nested records and unions.

For example, if a Course model has a union of Instructors and TeachingAssistants, a related resource can be added to the Instructor model through the union.

This also changes the variable interpolation to support referencing these nested models too.