Better handle reverse relations on empty lists (#162)
When lists were stringified and combined for reverse relations, we could end up with a query parameter that looks like `”,idTwo”` in cases where one element’s list of ids was empty. This causes the request to be invalid, and for data to not be returned.

This fix simply filters out empty strings, to ensure that all ids are valid when stringifying the args.