Use pegasus schemas for nested courier models inside case classes (#163)
In some cases, we have a scala case class that has a nested Courier record / union as a field inside of it. Currently, when we try to infer the schema, we'll come across the nested union schema, and mark it as "unable to be inferred". However, we have the schema available at that point, we're just not using it.

This change checks every class we try to infer (including nested classes inside a case class) and check if it's a pegasus schema. If it is, we pull the known schema from that, and ignore all attempted inferred schemas.