Make nullable field retrieval safer
DataMap’s `getDataList` field can return a null value. However, we were immediately calling `.asScala` on a lot of these values, which would obviously break. To fix this, we wrap the possibly-null value in an `Option` and then handle that safely (by either falling back to an empty list or calling getOrElse(null) after only calling asScala on the unwrapped value.