Namespace all union and record types under a resource with the resource name
Reverse relations on models (including on nested models and unions) are declared inside the resource class instead of on the courier model. However, GraphQL schemas are normalized, so a union with the name `org.coursera.course.CourseUnion` will be extracted out to its own type, and is not scoped under the resource that is including it at all.

This has the potential to cause problems when two resources have a model that include a field with a shared union or record. If one resource defines a related resource on a nested union or record, and another resource doesn’t, then there will be two different representations of that nested type from the two merged models. When these are generated into a GraphQL schema though, they will have the same name, and one schema will overwrite the other.

In order to solve this conflict, I’m choosing to namespace all of the records under the resource name. This will mean that the same union class will be represented with two different types in the GraphQL schema if it is used in two different resources.