Add DataSchemas for parameter types on naptime handlers (#138)
This change adds a new `typeSchema` field to the Parameter, which contains a DataSchema (serialized as a DataMap) that defines the parameter type. This allows the type of a parameter to be serialized in a better way than a string, so it can be parsed by the GraphQL assembler and used in schema generation. (Presently, the scala typename is being parsed with a regex, but that’s terrible)

While GraphQL has the ability to support default values in its schema, it turns out that it isn’t actually necessary to worry about default values in GraphQL. If a request from GraphQL goes to the underlying naptime resource without a parameter, the naptime resource will fill in the default itself.

This saves us the trouble of figuring out how to serialize default values without json / courier formats available.