Support typedDefinitions in unions
Many of our Courier unions use `typedDefinitions`, which allows providing a “friendly name” to a union type, instead of using the fully qualified name.

This change adds supports for these typedDefinitions to Courier, both by changing the name of the member key in the union type, as well as parsing the typedDefinition response correctly.