Replace double underscores (#187)
Fields that begin with two underscores are reserved in GraphQL. However, some of our downstream fields also have names that begin with two underscores. In order to work around this GraphQL spec requirement, we simply replace the two underscores with a single underscore for our field names in the schema.