Move logic for resolving typedDefinition unions
The logic for handling typedDefinition unions was in the field that was built for a union. However, in cases where unions with typedDefinitions were subtypes inside a list, the logic in the field resolver was never being run.

This change extracts the typedDefinition logic to the `typeOf` method on the union and the resolver on the subfield in the union, to ensure that typedDefinition unions work inside lists too.