Initial work of GraphQL Engine Rewrite (#202)
Sangria is already optimized for handling fetching data, and it's much easier to reason about data when you're fetching it in the GraphQL ecosystem instead of as a pre-process.

This diff removes the Naptime engine, and uses Sangria's DeferredResolvers to batch and fetch data in the GraphQL query. This also should fix a lot of bugs around fetching related resource data, supporting `@skip` directives, better supporting typedDefinitions, and other unknown bugs that we haven't found yet.

We also add an extension that returns the original url used to fetch each field, which should help further with debugging.