Add `debugMode` flag and only send sourceUrls if in debug mode (#207)
We're currently sending the sourceUrl extension on every response. However, this ends up being a very large response that is for debugging purposes only. This change adds a debugMode to the context, and we'll only send the sourceUrl data if the debug mode is enabled (it's disabled by default). Filters will be able to modify this flag, to allow for custom headers / authorization checking to enable debug mode.